renamed helpers -> logic in main src file since this contains all the logic in the webapp, did some minor refactoring for server too
