+++ mobile-browser-based-version/server/src/logic/federated/request_handlers.ts
@@ -4,10 +4,10 @@ import msgpack from 'msgpack-lite'
 import * as config from '../../server.config'
 import {
   averageWeights
-} from '../../helpers/tensor_operations'
+} from './tensor_helpers/tensor_operations'
 import {
   assignWeightsToModel
-} from '../../helpers/tensor_serializer'
+} from './tensor_helpers/tensor_serializer'
 import { getTasks } from '../../tasks/tasks_io'
 import { AsyncWeightsHolder } from './async_weights_holder'
 import * as tf from '@tensorflow/tfjs'
