+++ mobile-browser-based-version/server/src/logic/federated/tensor_helpers/tensor_serializer.ts
