+++ mobile-browser-based-version/server/src/router/authenticator.ts
