+++ mobile-browser-based-version/server/src/router/router.ts
@@ -3,7 +3,7 @@ import _ from 'lodash'
 import * as handlers from '../logic/federated/request_handlers'
 import { writeNewTask, getTasks } from '../tasks/tasks_io'
 import { ExpressPeerServer } from 'peer'
-import { makeID } from '../helpers/authenticator'
+import { makeID } from './authenticator'
 import { createProxyMiddleware } from 'http-proxy-middleware'
 import * as config from '../server.config'
