--- mobile-browser-based-version/server/src/test/helpers.ts
@@ -1,3 +1,0 @@
-export function sleep (time) {
-  return new Promise((resolve) => setTimeout(resolve, time))
-}
