+++ mobile-browser-based-version/src/components/building_frames/csv/CsvTrainingFrame.vue
@@ -116,7 +116,7 @@
 import TrainingFrame from '../containers/TrainingFrame.vue'
 import IconCard from '../../containers/IconCard.vue'
 import Bezier2 from '../../../assets/svg/Bezier2.vue'
-import { CsvTaskHelper } from '@/helpers/task_definition/csv/helper'
+import { CsvTaskHelper } from '../../../logic/task_definition/csv/csv_task_helper'
 
 export default {
   name: 'csv-training-frame',
