+++ mobile-browser-based-version/src/components/building_frames/upload/preview/data_types/FileData.vue
@@ -13,7 +13,7 @@
 </template>
 
 <script>
-import { FileUploadManager } from '../../../../../helpers/data_validation/file_upload_manager'
+import { FileUploadManager } from '../../../../../logic/data_validation/file_upload_manager'
 import ElementData from './ElementData.vue'
 
 export default {
