+++ mobile-browser-based-version/src/components/building_frames/upload/preview/data_types/ImageData.vue
@@ -14,7 +14,7 @@
   </element-data>
 </template>
 <script>
-import { FileUploadManager } from '../../../../../helpers/data_validation/file_upload_manager'
+import { FileUploadManager } from '../../../../../logic/data_validation/file_upload_manager'
 import ElementData from './ElementData.vue'
 
 export default {
