+++ mobile-browser-based-version/src/logic/communication/authenticator.ts
