+++ mobile-browser-based-version/src/logic/communication/client.ts
