+++ mobile-browser-based-version/src/logic/communication/decentralised/check_buffer.ts
