+++ mobile-browser-based-version/src/logic/communication/decentralised/communication_codes.ts
