+++ mobile-browser-based-version/src/logic/communication/tensor_serializer.ts
