+++ mobile-browser-based-version/src/logic/data_validation/file_upload_manager.ts
