+++ mobile-browser-based-version/src/logic/data_validation/helpers_image_tasks.ts
