+++ mobile-browser-based-version/src/logic/logging/console_logger.ts
