+++ mobile-browser-based-version/src/logic/logging/logger.ts
