+++ mobile-browser-based-version/src/logic/memory/model_io.ts
