+++ mobile-browser-based-version/src/logic/task_definition/csv/csv_task.ts
@@ -1,6 +1,6 @@
 import * as d3 from 'd3'
 import * as tf from '@tensorflow/tfjs'
-import { checkData } from '../../data_validation/helpers_csv_tasks'
+import { checkData } from '../../data_validation/check_data'
 import { Task } from '../base/task'
 
 type HeaderElem = String | Number;
