+++ mobile-browser-based-version/src/logic/task_definition/image/image_task.ts
