+++ mobile-browser-based-version/src/logic/testing/tester.ts
