+++ mobile-browser-based-version/src/logic/training/decentralised/training_informant.ts
