+++ mobile-browser-based-version/src/logic/training/federated/training_informant.ts
