+++ mobile-browser-based-version/src/logic/training/trainer.ts
