+++ mobile-browser-based-version/src/logic/training/training_manager.ts
